Bueno strjoin

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: apico-su <apico-su@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/17 18:34:50 by apico-su          #+#    #+#             */
-/*   Updated: 2021/04/16 16:49:24 by apico-su         ###   ########.fr       */
+/*   Updated: 2021/04/19 16:07:26 by apico-su         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int	ft_atoi(const char *str)
 {
-	int	numero;
-	int	negativo;
-	int	i;
+	long long	numero;
+	int			negativo;
+	int			i;
 
 	numero = 0;
 	negativo = 1;
@@ -36,6 +36,6 @@ int	ft_atoi(const char *str)
 	else if (str[i] == '+')
 		i++;
 	while (str[i] >= '0' && str[i] <= '9')
-		numero = numero * 10 + str[i++] - 48;
-	return (numero * negativo);
+		numero = numero * 10 + (long long)str[i++] - 48;
+	return ((int)numero * negativo);
 }
